### 2021-10-12 #3819 @whedon
[PRE REVIEW]: ordPens: An R package for Selection, Smoothing and Principal Components Analysis for Ordinal Variables
**Submitting author:** @ahoshiyar (<a href="http://orcid.org/0000-0002-5702-130X">Aisouda Hoshiyar</a>)
**Repository:** <a href="https://github.com/ahoshiyar/ordPens" target ="_blank">https://github.com/ahoshiyar/ordPens</a>
**Version:** 1.0.0
