### 2021-10-14 cf6305d
JOSS reference formatting
This relates to pre-review issue https://github.com/openjournals/joss-reviews/issues/3819. I have given one example of how to fix refences so they are in the right format for a JOSS paper.
