+++ paper/references.bib
@@ -22,7 +22,8 @@
   journal = {Electronic Journal of Statistics},
   volume = {5},
   pages = {1935--1959},
-  year = {2011}
+  year = {2011},
+  doi = {10.1214/11-EJS661}
 }
 
 @article{GerTut:2009,
