+++ references.bib
@@ -1,0 +1,453 @@
+% Encoding: UTF-8
+@manual{Falissad:2009,
+    title = {psy: Various procedures used in psychometry},
+    author = {B. Falissard},
+    year = {2009},
+    note = {R package version 1.0}
+  }
+
+@Article{GerHogObeTut:2011,
+  author  = {J. Gertheiss and S. Hogger and C. Oberhauser and G. Tutz},
+  title   = {Selection of Ordinally Scaled Independent Variables with Applications to International Classification of Functioning Core Sets},
+  journal = {Journal of the Royal Statistical Society C},
+  year    = {2011},
+  volume  = {60},
+  pages   = {377--395},
+  doi = {10.1111/j.1467-9876.2010.00753.x}
+}
+
+@article{GerOeh:2011,
+  title = {Testing linearity and relevance of ordinal predictors},
+  author = {Gertheiss, J. and Oehrlein, F.},
+  journal = {Electronic Journal of Statistics},
+  volume = {5},
+  pages = {1935--1959},
+  year = {2011},
+  doi={10.1214/11-ejs661}
+}
+
+@article{GerTut:2009,
+    author   = {J. Gertheiss and G. Tutz} ,
+    title    = {Penalized Regression with Ordinal Predictors},
+    journal  = {International Statistical Review},
+    year     = {2009},
+    volume   = {77},
+    pages = {345--365},
+    doi={10.1111/j.1751-5823.2009.00088.x}
+}
+
+@article{GerTut:2010,
+  author =  {J. Gertheiss and G. Tutz},
+  title =   {Sparse Modeling of Categorial Explanatory Variables},
+  year =    2010,
+  journal = {The Annals of Applied Statistics},
+  volume = {4},
+  pages = {2150--2180},
+  doi={10.1214/10-aoas355}
+}
+
+@article{Gertheiss:2014,
+  title={ANOVA for factors with ordered levels},
+  author={Gertheiss, J.},
+  journal={Journal of Agricultural, Biological, and Environmental Statistics},
+  volume={19},
+  number={2},
+  pages={258--277},
+  year={2014},
+  publisher={Springer},
+  doi={10.1007/s13253-014-0170-5}
+}
+
+@article{Sweeney:2016,
+  title={Testing differentially expressed genes in dose-response studies and with ordinal phenotypes},
+  author={Sweeney, E. and Crainiceanu, C. and Gertheiss, J.},
+  journal={Statistical applications in genetics and molecular biology},
+  volume={15},
+  number={3},
+  pages={213--235},
+  year={2016},
+  publisher={De Gruyter},
+  doi={10.1515/sagmb-2015-0091}
+}
+
+@Article{JouEtal:1988,
+  author  = {R. Jouvent and C. Vindreau and M. Montreuil and C. Bungender and D. Windlocher},
+  title   = {La clinique polydimensionnelle de humeur depressive: {N}ouvelle version echelle {EHD}},
+  journal = {Psychiatrie et Psychobiologie},
+  year    = {1988},
+  volume  = {3},
+  pages   = {245--254},
+}
+
+@Article{LinMeuKooGro:2007,
+  author  = {M. Linting and J. J. Meulman and A. J. {van der Kooij} and P. J. F. Groenen},
+  title   = {Nonlinear Principal Components Analysis: Introduction and Application},
+  journal = {Psychological Methods},
+  year    = {2007},
+  volume  = {12},
+  pages   = {336--358},
+}
+
+@Manual{R,
+  title        = {R: A Language and Environment for Statistical Computing},
+  author       = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
[…]
