### 2021-11-04 #3828 @fartist
I am reviewing it right now. Sent from Mail for Windows From: Franjo IvankovicSent: Wednesday, November 3, 2021 9:37 PMTo: openjournals/joss-reviewsCc: Feeeeeeeeeeng; MentionSubject: Re: [openjournals/joss-reviews] [REVIEW]: ordPens: An R package for Selection, Smoothing and Principal Components Analysis for Ordinal Variables (#3828) I am still waiting on @ahoshiyar to address the open issues.—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or unsubscribe.Triage notifications on the go with GitHub Mobile for iOS or Android.
