+++ Contributing.md
@@ -1,0 +1,57 @@
+---
+output:
+  html_document: default
+  pdf_document: default
+---
+# How to contribute to `ordPens`
+
+Thank you for taking the time to contribute to the development of `ordPens`!
+Before you start:
+
+* It is important to have a valid [GitHub account](https://github.com/signup/free).
+
+### Did you find a bug? 
+ 
+* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/ahoshiyar/ordPens/issues).
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/ahoshiyar/ordPens/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible, and a **reproducible code sample** demonstrating the expected behavior that is not occurring.
+
+ 
+ 
+### Did you write a patch that fixes a bug?
+
+* Open a new GitHub pull request with the patch.
+
+* Ensure the PR description clearly describes the problem and solution. Include the relevant issue number if applicable.
+ 
+ 
+ 
+### Do you intend to add a new feature or change an existing one?
+ 
+* Start your work on your fork of the repository.  
+
+* Install all development dependences with `devtools::install_dev_deps()`, and then make sure the package passes R CMD check by running `devtools::check()`. 
+
+* Create a Git branch for your pull request (PR).  
+
+* Commit messages should be descriptive, mentioning what was changed and why, and also reference the relevant issue number. 
+  
+* Commit to git, and then send a PR to [ordPens](https://github.com/ahoshiyar/ordPens/pulls) with a clear list of what you've done.
+
+### Copyright issues
+ 
+* On submission, it is crucial your PR includes the following statement: You own the copyright on the code being contributed, and you hereby grant `ordPens` repo cph unlimited license to use this code in this version or any future version of `ordPens`. You reserve all other rights to the code.
+
+* The Pull Requests are thereafter reviewed, with feedbacks communicated as soon as possible.
+
+
+
+### Do you have questions about the source code?
+
+* Ask any question about how to use `ordPens` via E-mail to the maintainer.
+
+ 
+ 
+Thanks!  
+
+ 
