### 2021-11-04 459d40f
add installation & community guidelines to README
