+++ README.Rmd
@@ -35,6 +35,38 @@ Also, performance evaluation and selection of an optimal penalty parameter provi
 * `ordFusion` fits dummy coefficients of ordinally scaled independent variables with a fused lasso penalty for fusion and selection.
 * A new type of spline basis for ordered factors `s(..., bs = "ordinal")`is provided, such that smooth terms in the `mgcv::gam()` formula can be used as an alternative and extension to `ordSmooth()`. Additionally, generic functions for prediction and plotting are provided.  
  
+## Installation & getting started
+
+For standard use, install `ordPens` from [CRAN](https://cran.r-project.org/package=ordPens):
+```{r CRAN-install, eval=FALSE}
+install.packages("ordPens")
+```
+
+The development version of the package may be installed from GitHub:
+```{r git-install, eval=FALSE} 
+devtools::install_git("https://github.com/ahoshiyar/ordPens", build_vignettes = TRUE)
+```
+
+For a detailed overview about the functionalities and given examples type:
+```{r vignettes, eval=FALSE} 
+library(ordPens)
+vignette("ordPens", package = "ordPens")
+vignette("ordPCA", package = "ordPens")
+```
+
+## Issues
+
+If you encounter any bugs or have any specific feature requests, please [file an issue](https://github.com/ahoshiyar/ordPens/issues).
+
+## Contributions & Code of conduct 
+
+Contributions are very welcome. Interested contributors should consult the
+[contribution
+guidelines](https://github.com/ahoshiyar/ordPens/blob/master/Contributing.md)
+prior to submitting a pull request.
+  
+Please note that the `ordPens` package is released with a [Contributor Code of Conduct](https://www.contributor-covenant.org/version/2/0/code_of_conduct/). By contributing to this project, you agree to abide by its terms.
+ 
 
 ## References
 
