### 2021-11-04 7d76446
add installation & community guidelines to README
