+++ paper/references.bib
@@ -449,5 +449,18 @@
   }
 
 
+@article{Huang:2021,
+  doi = {10.21105/joss.03634},
+  url = {https://doi.org/10.21105/joss.03634},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {65},
+  pages = {3634},
+  author = {Mingze Huang and Christian L. Müller and Irina Gaynanova},
+  title = {latentcor: An R Package for estimating latent correlations from mixed data types},
+  journal = {Journal of Open Source Software}
+}
+
 
 @Comment{jabref-meta: databaseType:bibtex;}
