### 2021-11-04 f24499a
Adding Possible related methods to paper.md
