### 2021-11-11 ahoshiyar/ordPens#9 @fartist
[REVIEW]reference details
1.
 Huang, M., Müller, C. L., & Gaynanova, I. (2021). Latentcor: An r package for estimating
114 latent correlations from mixed data types. Journal of Open Source Software, 6(65), 3634.

"r" should be in uppercase.

2. line 7, add a couple of references for general statistical treatment for using regularization for ordinal variables.

3. line 70, why the "thresholds" is denoted as "latent"? they are just statistical parameters, I would suggest taking "(latent)" here to avoid confusion with the latent continuous variable.
