### 2021-11-12 ahoshiyar/ordPens#9 @osorensen
Thanks for pointing this out @fartist. A comment to @ahoshiyar is that the reference formatting can be fixed by adding curly braces {} around the R in the BibTeX reference, so it becomes {R}. Then it will appear as an uppercase R.
