### 2021-11-12 ahoshiyar/ordPens#10 @mingzehuang
Tests
Your package runs very smooth and I realize your method has an advantage on speed performance since your underline math method is less computationally complex than some other methods! I think I only have the last issue left if you could address it. The checklist asks "automated tests". It seems that you don't make "tests" folder to have some automated tests. This can be done by `usethis::use_testthat()`. [See here](https://r-pkgs.org/tests.html) Although you don't have to have very high test coverage since JOSS doesn't require a specific level.
