### 2021-11-12 #3828 @mingzehuang
Great work @ahoshiyar ! My last comment is to generate automated tests which is required by the checklist. [Tests](https://github.com/ahoshiyar/ordPens/issues/10)
