+++ paper/references.bib
@@ -458,7 +458,7 @@
   number = {65},
   pages = {3634},
   author = {Mingze Huang and Christian L. Müller and Irina Gaynanova},
-  title = {latentcor: An R Package for estimating latent correlations from mixed data types},
+  title = {latentcor: An {R} Package for estimating latent correlations from mixed data types},
   journal = {Journal of Open Source Software}
 }
 
