### 2021-11-18 ahoshiyar/ordPens#9 @ahoshiyar
Thank you!
1. Thanks, has been changed.
2. These are mentioned in line 21, but I added them also to line 7 now.
3. Thank you for pointing out, the "latent" has been removed.
