+++ tests/testthat/test-penalized-estimation.R
@@ -1,0 +1,344 @@
+
+library(testthat)
+library(ordPens)
+
+
+context("check ordinal Smoothing/Selection/Fusion functions")
+ 
+
+## checks on warning and error messages
+test_that("ordSmooth, ordSelect, ordFusion work properly",
+          {
+            
+            data(ICFCoreSetCWP) 
+            x <- ICFCoreSetCWP[, 1:67] + matrix(c(rep(1, 50), rep(5, 16), 1),
+                                                nrow(ICFCoreSetCWP), 67,
+                                                byrow = TRUE)
+            y <- ICFCoreSetCWP$phcs
+            lambda <- c(1000, 500, 200, 100, 70, 50, 30, 20, 10, 1) 
+            
+            x <- rbind(x, rep(1,67))
+            x <- rbind(x, c(rep(5, 50), rep(9,16), 5))
+            y <- c(y, NA, NA)
+            
+            # checks on x 
+            expect_error(  
+              ordSmooth(x = list(x), y = y, lambda = lambda),
+              "x has to be a matrix, numeric vector or data.frame" 
+            ) 
+            expect_error(  
+              ordSelect(x = list(x), y = y, lambda = lambda),
+              "x has to be a matrix, numeric vector or data.frame" 
+            ) 
+            expect_error(  
+              ordFusion(x = list(x), y = y, lambda = lambda),
+              "x has to be a matrix, numeric vector or data.frame" 
+            )
+            
+            x.na <- x
+            x.na[1] <- NA
+            
+            expect_error(   
+              ordSmooth(x = x.na, y = y, lambda = lambda),
+              "Missing values in x are not allowed" 
+            ) 
+            expect_error(   
+              ordSelect(x = x.na, y = y, lambda = lambda),
+              "Missing values in x are not allowed"
+            ) 
+            expect_error(   
+              ordFusion(x = x.na, y = y, lambda = lambda),
+              "Missing values in x are not allowed"
+            ) 
+            
+            # checks on response y
+            expect_error(  
+              ordSmooth(x = x, y = list(y), lambda = lambda),
+              "y has to be numeric"
+            ) 
+            expect_error(  
+              ordSelect(x = x, y = list(y), lambda = lambda),
+              "y has to be numeric"
+            ) 
+            expect_error(  
+              ordFusion(x = x, y = list(y), lambda = lambda),
+              "y has to be numeric"
+            )
+            
+            expect_error(  
+              ordSmooth(x = x, y = y, lambda = lambda, model = "logit"),
+              "y has to be 0/1 coded"
+            ) 
+            expect_error(  
+              ordSelect(x = x, y = y, lambda = lambda, model = "logit"),
+              "y has to be 0/1 coded"
+            ) 
+            expect_error(  
+              ordFusion(x = x, y = y, lambda = lambda, model = "logit"),
+              "y has to be 0/1 coded"
+            )
+  
+            expect_error(  
+              ordSmooth(x = x, y = y, lambda = lambda, model = "poisson"),
+              "y has to contain nonnegative integers"
+            ) 
+            expect_error(  
+              ordSelect(x = x, y = y, lambda = lambda, model = "poisson"),
+              "y has to contain nonnegative integers"
+            ) 
+            expect_error(  
+              ordFusion(x = x, y = y, lambda = lambda, model = "poisson"),
+              "y has to contain nonnegative integers"
+            )
+            
+            y.neg <- round(y) 
+            y.neg[1] <- -1
+            
+            expect_error(  
+              ordSmooth(x = x, y = y.neg, lambda = lambda, model = "poisson"),
+              "y has to contain nonnegative integers"
+            ) 
+            expect_error(  
+              ordSelect(x = x, y = y.neg, lambda = lambda, model = "poisson"),
[…]
