+++ DESCRIPTION
@@ -11,10 +11,11 @@ Depends: grplasso, mgcv, RLRsim, quadprog, glmpath
 Suggests: 
   psy,
   knitr,
-  rmarkdown 
+  rmarkdown,
+  testthat (>= 3.0.0)
 VignetteBuilder: knitr
 License: GPL-2
 LazyLoad: yes
 NeedsCompilation: no 
 Repository: CRAN   
-
+Config/testthat/edition: 3
