### 2021-11-19 ahoshiyar/ordPens#9 @ahoshiyar
Thank you @fartist ! :)

Thanks, has been changed.
These are mentioned in line 21, but I added them also to line 7 now.
Thank you for pointing out, the "latent" has been removed.
