### 2021-11-19 ahoshiyar/ordPens#11 @osorensen
Editorial remarks on JOSS submission
Hello @ahoshiyar, I'm the editor of your submission in [this repository](https://github.com/openjournals/joss-reviews/issues/3828), and have some comments on the manuscript for which I'll create some issues in this repository.

My first comment is regarding the following statement, in **Summary**

> The main idea behind smoothing in the regression context is to maximize the penalized log-likelihood by introducing a penalty term and a tuning parameter controlling the amount of penalty.

I'm not sure if this is technically correct. Taken literally, the sentence claims that the log-likelihood is maximized by introducing a penalty term and a tuning parameter.

What about rewriting to something like this: _In the regression context, smoothing is obtained by introducing a penalty term and a tuning parameter controling the amount of penalty. Adding the penalty term to the likelihood function yields the penalized likelihood, which is then maximized._
