### 2021-11-19 ahoshiyar/ordPens#12 @osorensen
Imprecise statement?
Here is another comment on the submission. In the **Summary**, you write

> Smoothing only can be done by penalizing the sum of squared differences of adjacent coefficients.

Isn't this only correct if (1) there is only a single only variable in the model, and (2) the indicators for the different levels are properly ordered. I'm not an expert on regression with ordinal variables, but perhaps something like this would be more precise?

_Smoothing only can be done by penalizing the sum of squared differences of adjacent coefficients for a given variable, subject to proper ordering._

This issue relates to https://github.com/openjournals/joss-reviews/issues/3828
