### 2021-11-19 ahoshiyar/ordPens#13 @osorensen
Add a comma?
We're down to the details here now, but on lines 43-44 you write

> Furthermore, testing for differences in the means, known as analysis of variance (ANOVA) is provided for ordered factors by the function `ordAOV()` penalizing (squared) differences of adjacent means.

I think there should be a comma after (ANOVA), yielding: _Furthermore, testing for differences in the means, known as analysis of variance (ANOVA), is provided for ordered factors by the function `ordAOV()` penalizing (squared) differences of adjacent means._

This issue relates to https://github.com/openjournals/joss-reviews/issues/3828
