### 2021-11-19 ahoshiyar/ordPens#14 @osorensen
Rewrite?
On line 88, you write

> ...; type also `?ICFCoreSetCWP`.

I think the sentence becomes more correct if you change from "type also" to "or by typing", so it becomes: _"...; or by typing `?ICFCoreSetCWP`"_

This issue relates to https://github.com/openjournals/joss-reviews/issues/3828
