### 2021-11-19 ahoshiyar/ordPens#15 @osorensen
Figure 1 not referenced in text
Please add a reference to Figure 1 somewhere in the text. The best place to put it is probably just above the block of code, starting at line 90. It is sufficient to write something along the lines of _The output is shown in Figure 1._

This issue relates to https://github.com/openjournals/joss-reviews/issues/3828
