+++ paper/paper.md
@@ -52,7 +52,7 @@ $$
 $$ 
 For more information on the original group lasso (for nominal predictors and grouped variables in general), see @Meier:2008 and @Yuan:2006. For details on the fused lasso, see @Tibshirani:2005.
 In the case of smoothing only, the package includes auxiliary functions such that ``mgcv::gam()`` [@Wood:2008; @Wood:2017] can be used for fitting generalized linear and additive models with first- and second-order ordinal smoothing penalty as well as built-in smoothing parameter selection. Also, ``mgcv`` tools for further statistical inference can be used, see @Gertheiss:2021 for details.
-Furthermore, testing for differences in the means, known as analysis of variance (ANOVA) is provided for ordered factors by the function `ordAOV()` penalizing (squared) differences of adjacent means. Testing for differentially expressed genes, when analyzing microarrays of gene expression data, is incorporated by the function `ordGene()`. 
+Furthermore, testing for differences in the means, known as analysis of variance (ANOVA), is provided for ordered factors by the function `ordAOV()` penalizing (squared) differences of adjacent means. Testing for differentially expressed genes, when analyzing microarrays of gene expression data, is incorporated by the function `ordGene()`. 
 Technical details can be viewed from @Gertheiss:2014 and @Sweeney:2016, respectively.
 
 
@@ -76,8 +76,8 @@ For penalized regression and ordinal ANOVA see also ``vignette("ordPens", packag
   
 # ``ordPens`` in action
 
-This example illustrates penalized non-linear PCA on the so-called brief ICF core set on Chronic Widespread Pain (CWP) consisting of 26 ordinally scaled variables. Details on the data can be found in @GerHogObeTut:2011; type also `?ICFCoreSetCWP`.
-Analysis of the so-called comprehensive core set for CWP, consisting of 67 ICF variables, is found in @Hoshiyar:2021. Note that the `ordPCA` vignette also analyzes the comprehensive core set. 
+This example illustrates penalized non-linear PCA on the so-called brief ICF core set on Chronic Widespread Pain (CWP) consisting of 26 ordinally scaled variables. Details on the data can be found in @GerHogObeTut:2011; or by typing `?ICFCoreSetCWP`.
+Analysis of the so-called comprehensive core set for CWP, consisting of 67 ICF variables, is found in @Hoshiyar:2021. Note that the `ordPCA` vignette also analyzes the comprehensive core set. Figure 1, generated by the following code, illustrates the estimated coefficients of selected variables for different values of the penalty parameter $\lambda$ along with cross-validation results.  
 ```r
 
 library(ordPens)
