### 2021-11-27 #3828 @Kevin-Mattheus-Moerman
@ahoshiyar I have proofread your paper which seems in order. However the ZENODO archive needs changes as listed below:

- [ ] Please amend the license on the ZENODO archive (you have to do this manually) to match your software license: https://github.com/ahoshiyar/ordPens/blob/master/LICENSE, it currently lists cc-by.
