### 2021-11-29 #3828 @ahoshiyar
Thank you @Kevin-Mattheus-Moerman, I see..
However, Zenodo seems to provide only the Creative Commons Attribution licenses. I've found also other JOSS related packages having different repo licenses and Zenodo licenses and was confused.
Could you please give me advice on what to do in this case?
