### 2021-12-03 #3828 @Kevin-Mattheus-Moerman
@arfon can you comment on this. Shouldn't the zenodo archive be editable to change the license to match that of the software?
