### 2021-12-05 #3828 @arfon
I think this may be a limitation of how Zenodo works. As long as the software has a license file in the repository (which it will) with an OSI approved license then I think this is ok
