### 2021-12-06 #3828 @Kevin-Mattheus-Moerman
@ahoshiyar congratulations on getting this work published in JOSS!

Thanks @fartist, @FranjoIM, @mingzehuang for you review efforts, and thank you @osorensen for editing this work!
